Rename QuestionFileList.vue to QuestionFileListWithoutVuex.vue
