+++ static/src/questionnaire-detail.js
@@ -1,11 +1,6 @@
 import '@babel/polyfill'
 import './utils/polyfills.js'
 
-import InfoBar from './utils/InfoBar'
-import QuestionBox from './questions/QuestionBox'
-import QuestionFileList from './questions/QuestionFileList'
-import ResponseDropzone from './questions/ResponseDropzone'
-import ResponseFileList from './questions/ResponseFileList'
 import QuestionnaireDetailPage from './questionnaires/QuestionnaireDetailPage'
 import Vue from 'vue/dist/vue.js'
 import Vuex, { mapActions } from 'vuex'
@@ -13,15 +8,11 @@ import { store } from './store'
 
 Vue.use(Vuex)
 
-let questionnaireDetailApp = new Vue({
+// eslint-disable-next-line no-new
+new Vue({
   store,
   el: '#questionnaire-detail-app',
   components: {
-    InfoBar,
-    QuestionBox,
-    QuestionFileList,
-    ResponseDropzone,
-    ResponseFileList,
     QuestionnaireDetailPage,
   },
   methods: {
